Move variable-differences testing into own testfile.
