+++ tests/unit/utils/compare_nc_datasets/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for :mod:`ncdata.utils._compare_nc_datasets`."""
