+++ tests/unit/utils/compare_nc_datasets/test_dataset_differences__mainfunctions.py
@@ -6,7 +6,6 @@ Split in two files ...
     * ( ALSO: "additional" tests (q.v.) cover subsidiary routines and the
         main API usage modes. )
 """
-import dask.array as da
 import numpy as np
 import pytest
 
@@ -405,336 +404,6 @@ class TestCompareVariables__metadata:
         expected = [
             f"{self.location_string} variable lists do not match: "
             "['v1', 'v2'] != ['v2']"
-        ]
-        check(errs, expected)
-
-    def test_var_dims__reorder(self, order_checking):
-        # N.B. here we check behaviour of the DIMENSIONS order control, but this does
-        # not apply to dimensions order in a variable,which is *always* significant.
-        self.vars["v1"].dimensions = self.vars["v1"].dimensions[::-1]
-        # N.B. the data shape doesn't now correspond, but that won't matter as, with
-        #  mismatched dimensions, the data won't be checked.
-
-        do_orderchecks = decode_ordercheck(order_checking)
-        errs = dataset_differences(
-            self.data1, self.data2, check_dims_order=do_orderchecks
-        )
-
-        expected = [
-            f'{self.location_string} variable "v1" dimensions differ : '
-            "('y', 'x') != ('x', 'y')"
-        ]
-        check(errs, expected)
-
-    def test_var_dims__extra_or_missing(self, order_checking):
-        # N.B. here we check for DIMENSIONS order check control.
-        self.vars["v1"].dimensions = self.vars["v1"].dimensions[:-1]
-        # N.B. the data shape doesn't now correspond, but that won't matter as, with
-        #  mismatched dimensions, the data won't be checked.
-
-        do_orderchecks = decode_ordercheck(order_checking)
-        errs = dataset_differences(
-            self.data1, self.data2, check_dims_order=do_orderchecks
-        )
-
-        expected = [
-            f'{self.location_string} variable "v1" dimensions differ : '
-            "('y', 'x') != ('y',)"
-        ]
-        check(errs, expected)
-
-
-class TestCompareVariables__dtype:
-    # Note: testing variable comparison via the 'main' public API instead of
-    #  via '_variable_differences'.  This makes sense because it is only called
-    #  in one way, from one place.
-    @staticmethod
-    def _vars_testdata():
-        def data():
-            return np.zeros(3)
-
-        testdata = NcData(
-            name="dataset_1",
-            dimensions=[NcDimension("x", 3)],
-            variables=[
-                NcVariable("v1", ("x"), data=data()),
-            ],
-        )
-        return testdata
-
-    @pytest.fixture(autouse=True)
-    def _vars_data(self):
-        self.data1, self.data2 = [self._vars_testdata() for _ in range(2)]
-        self.reference_var, self.testvar = (
-            ds.variables["v1"] for ds in (self.data1, self.data2)
-        )
-
-    def test_numbers_v_strings(self):
-        # Set a different dtype
-        # NB this is different from the actual data array, but that doesn't
-        #  matter, as it won't attempt to compare strings with numbers
-        self.testvar.dtype = np.dtype("S5")
-
-        # Test the comparison
-        errs = dataset_differences(self.data1, self.data2)
-        expected = [
-            'Dataset variable "v1" datatypes differ : '
-            "dtype('float64') != dtype('S5')"
-        ]
-        check(errs, expected)
-
-    @pytest.mark.parametrize("equaldata", [False, True])
-    def test_ints_v_floats(self, equaldata):
-        # In this case, there is also a data comparison to check.
-        v1 = self.testvar
-
-        new_dtype = np.dtype(np.int32)
-        v1.data = v1.data.astype(new_dtype)
-        if not equaldata:
-            v1.data.flat[0] += 1
-        v1.dtype = new_dtype
-
-        # Test the comparison
-        errs = dataset_differences(self.data1, self.data2)
-
-        expected = [
-            'Dataset variable "v1" datatypes differ : '
-            "dtype('float64') != dtype('int32')"
-        ]
-        if not equaldata:
-            expected.append(
-                'Dataset variable "v1" data contents differ, at 1 points: '
-                "@INDICES[(0,)] : LHS=[0.0], RHS=[1]"
-            )
-        check(errs, expected)
-
-    @pytest.mark.parametrize("equaldata", [False, True])
-    def test_wordlengths(self, equaldata):
-        # Test floats with wordlength difference -- assume ints are the same
-        # In this case, there is also a data comparison to check.
-        v1 = self.testvar
-
-        new_dtype = np.dtype(np.float32)
-        v1.data = v1.data.astype(new_dtype)
-        if not equaldata:
-            v1.data.flat[0] += 1
-        v1.dtype = new_dtype
-
-        # Test the comparison
-        errs = dataset_differences(self.data1, self.data2)
-
-        expected = [
-            'Dataset variable "v1" datatypes differ : '
-            "dtype('float64') != dtype('float32')"
-        ]
-        if not equaldata:
-            expected.append(
-                'Dataset variable "v1" data contents differ, at 1 points: '
-                "@INDICES[(0,)] : LHS=[0.0], RHS=[1.0]"
-            )
-        check(errs, expected)
-
-    @pytest.mark.parametrize("equaldata", [False, True])
-    def test_signed_unsigned(self, equaldata):
-        # Test floats with wordlength difference -- assume ints are the same
-        # In this case, there is also a data comparison to check.
-        new_dtype = np.dtype(np.int64)
-        v0 = self.reference_var
-        v0.data = v0.data.astype(new_dtype)
-        v0.dtype = new_dtype
-
-        new_dtype = np.dtype(np.uint64)
-        v1 = self.testvar
-        v1.data = v1.data.astype(new_dtype)
-        if not equaldata:
-            v1.data.flat[0] += 1
-        v1.dtype = new_dtype
-
-        # Test the comparison
-        errs = dataset_differences(self.data1, self.data2)
-
-        expected = [
-            'Dataset variable "v1" datatypes differ : '
-            "dtype('int64') != dtype('uint64')"
-        ]
-        if not equaldata:
-            expected.append(
-                'Dataset variable "v1" data contents differ, at 1 points: '
-                "@INDICES[(0,)] : LHS=[0], RHS=[1]"
-            )
-        check(errs, expected)
-
-
-class TestCompareVariables__data__checkcontrols:
-    # Note: testing variable comparison via the 'main' public API instead of
-    #  via '_variable_differences'.  This makes sense because it is only called
-    #  in one way, from one place.
-    @staticmethod
-    def _vars_testdata():
-        def data():
-            return np.arange(6.0).reshape((2, 3))
-
-        testdata = NcData(
-            name="dataset_1",
-            dimensions=[NcDimension("y", 2), NcDimension("x", 3)],
-            variables=[
-                NcVariable("v1", ("x"), data=data()),
-            ],
-        )
-        return testdata
-
-    @pytest.fixture(autouse=True)
-    def _vars_data(self):
-        self.data1, self.data2 = [self._vars_testdata() for _ in range(2)]
-        self.reference_var = self.data1.variables["v1"]
-        self.testvar = self.data2.variables["v1"]
-
-    def test_no_values_check(self):
-        self.testvar.data += 1
-        errs = dataset_differences(
-            self.data1, self.data2, check_var_data=False
-        )
-        check(errs, [])
-
-    def test_print_bad_nprint(self):
-        msg = "'show_n_diffs' must be >=1 : got 0."
-        with pytest.raises(ValueError, match=msg):
-            dataset_differences(
-                self.data1, self.data2, show_n_first_different=0
-            )
-
-    @pytest.mark.parametrize("ndiffs", [1, 2, 3])
-    def test_ndiffs(self, ndiffs):
-        self.testvar.data.flat[1 : ndiffs + 1] += 1
-        errs = dataset_differences(self.data1, self.data2)
-        detail = {
-            1: "[(0, 1)] : LHS=[1.0], RHS=[2.0]",
-            2: "[(0, 1), (0, 2)] : LHS=[1.0, 2.0], RHS=[2.0, 3.0]",
-            3: (
-                "[(0, 1), (0, 2), ...] : "
-                "LHS=[1.0, 2.0, ...], RHS=[2.0, 3.0, ...]"
-            ),
-        }[ndiffs]
-        expected = [
-            f'Dataset variable "v1" data contents differ, at {ndiffs} points: '
-            f"@INDICES{detail}"
-        ]
-        check(errs, expected)
-
-    @pytest.mark.parametrize("nprint", [1, 2, 3])
-    def test_show_n_first_different(self, nprint):
-        self.testvar.data.flat[1:3] += 1
-        errs = dataset_differences(
-            self.data1, self.data2, show_n_first_different=nprint
-        )
-        detail = {
-            1: "[(0, 1), ...] : LHS=[1.0, ...], RHS=[2.0, ...]",
-            2: "[(0, 1), (0, 2)] : LHS=[1.0, 2.0], RHS=[2.0, 3.0]",
-            3: "[(0, 1), (0, 2)] : LHS=[1.0, 2.0], RHS=[2.0, 3.0]",
-        }[nprint]
-        expected = [
-            f'Dataset variable "v1" data contents differ, at 2 points: '
-            f"@INDICES{detail}"
-        ]
-        check(errs, expected)
-
-
-class TestCompareVariables__data__diffreports:
-    # Note: testing variable comparison via the 'main' public API instead of
-    #  via '_variable_differences'.  This makes sense because it is only called
-    #  in one way, from one place.
-    @staticmethod
-    def _vars_testdata():
-        def data():
-            return np.arange(4.0)
-
-        testdata = NcData(
-            name="dataset_1",
-            dimensions=[NcDimension("x", 4)],
-            variables=[
-                NcVariable("v1", ("x"), data=data()),
-            ],
-        )
-        return testdata
-
-    @pytest.fixture(autouse=True)
-    def _vars_data(self):
-        self.data1, self.data2 = [self._vars_testdata() for _ in range(2)]
-        self.reference_var = self.data1.variables["v1"]
-        self.testvar = self.data2.variables["v1"]
-
-    @pytest.mark.parametrize("datavalues", ["same", "different"])
-    @pytest.mark.parametrize("masks", ["onemasked", "bothmasked"])
-    def test_masked(self, datavalues, masks):
-        different = datavalues == "different"
-        bothmasked = masks == "bothmasked"
-        testvar = self.testvar
-        testvar.data = np.ma.masked_array(testvar.data)
-        if different:
-            testvar.data[1:2] += 1
-        testvar.data[1:2] = np.ma.masked
-        if bothmasked:
-            self.reference_var.data = np.ma.masked_array(
-                self.reference_var.data
-            )
-            self.reference_var.data[1:2] = np.ma.masked
-        errs = dataset_differences(self.data1, self.data2)
-        if bothmasked:
-            expected = []
-        else:
-            expected = [
-                'Dataset variable "v1" data contents differ, at 1 points: '
-                "@INDICES[(1,)] : LHS=[1.0], RHS=[masked]"
-            ]
-        check(errs, expected)
-
-    @pytest.mark.parametrize("nans", ["onenans", "bothnans"])
-    def test_nans(self, nans):
-        bothnans = nans == "bothnans"
-        self.testvar.data[1:2] = np.nan
-        if bothnans:
-            self.reference_var.data[1:2] = np.nan
-        errs = dataset_differences(self.data1, self.data2)
-        if bothnans:
-            expected = []
-        else:
-            expected = [
-                'Dataset variable "v1" data contents differ, at 1 points: '
-                "@INDICES[(1,)] : LHS=[1.0], RHS=[nan]"
-            ]
-        check(errs, expected)
-
-    def test_scalar(self):
-        # Check how a difference of scalar arrays is reported
-        for value, var in enumerate([self.reference_var, self.testvar]):
-            var.dimensions = ()
-            var.data = np.array(value, dtype=var.dtype)
-        errs = dataset_differences(self.data1, self.data2)
-        expected = [
-            'Dataset variable "v1" data contents differ, at 1 points: '
-            "@INDICES[(0,)] : LHS=[0.0], RHS=[1.0]"
-        ]
-        check(errs, expected)
-
-    @pytest.mark.parametrize(
-        "argtypes", ["real_real", "real_lazy", "lazy_lazy"]
-    )
-    def test_real_and_lazy(self, argtypes):
-        type1, type2 = argtypes[:4], argtypes[-4:]
-        # fix the testvar to create a difference
-        self.testvar.data[1:2] += 1
-        # setup vars with lazy/real data arrays
-        for arraytype, var in zip(
-            [type1, type2], [self.reference_var, self.testvar]
-        ):
-            if arraytype == "lazy":
-                var.data = da.from_array(var.data, chunks=-1)
-        # compare + check results
-        errs = dataset_differences(self.data1, self.data2)
-        # N.B. the result should be the same in all cases
-        expected = [
-            'Dataset variable "v1" data contents differ, at 1 points: '
-            "@INDICES[(1,)] : LHS=[1.0], RHS=[2.0]"
         ]
         check(errs, expected)
